// src/app2/apis/questionnaire/type/diligence.type.ts
export interface DiligenceUserType {
  associated_firms_count: number;
  referral_data?: any;
  is_default: boolean;
  userName: string;
  password?: any;
  confirmPassword?: any;
  newPassword?: any;
  id: number;
  type: number;
  firstName: string;
  lastName: string;
  fullName: string;
  firmInfo?: any;
  accessLevel: string;
  accessLevelID: number;
  isAdmin: boolean;
  isReadOnly: boolean;
  is_approver: boolean;
  active: boolean;
  country?: any;
  avatarURL?: any;
  title?: any;
  psk?: any;
  twoFactorEnabled: boolean;
  anonymous: boolean;
  bounced_at?: any;
  phone?: any;
  eucAccepted: boolean;
  group_by_intro: boolean;
  review_functionality_intro: boolean;
  discussEUCAccepted: boolean;
  isFirstLogin: boolean;
  firmAccessCount: number;
  conversionDateTime?: any;
  isFirstDD: boolean;
  isFirstTemplate: boolean;
  statusCode: number;
  fundId: number;
  lastTouchPoint?: any;
  owner_user_id?: any;
  owner_name?: any;
  contact_type_ids?: any;
  contact_types?: any;
  type_ids?: any;
  street_address_1?: any;
  street_address_2?: any;
  city?: any;
  state?: any;
  zipcode?: any;
  country_id: number;
  phone_1?: any;
  phone_2?: any;
  fax?: any;
  relationship_status_id: number;
  relationship_status?: any;
  associated_funds: any[];
  associated_programs: any[];
  associated_strategies: any[];
  lockoutEndDateUtc?: any;
  key?: any;
  firm_name: string;
  is_tracking: boolean;
  is_verified: boolean;
  is_owner: boolean;
  is_new_user: boolean;
  skip_tour: boolean;
  approved_by?: any;
  approved_at?: any;
  send_invitation: boolean;
  can_change_firm: boolean;
  last_updated_at?: any;
  api_key?: any;
  firmwide_role?: any;
  is_saml_enabled: boolean;
  platform_active: boolean;
  relationship_active: boolean;
}

export interface DiligenceType {
  manager_firms: any[];
  id: number;
  name: string;
  entity_type: string;
  entity_id: number;
  entity_name: string;
  fund_name: string;
  vehicle_name?: any;
  template_id: number;
  template_version: number;
  template_name: string;
  status: string;
  investorfirm_name: string;
  investorfirm_id: number;
  managerfirm_name: string;
  managerfirm_id: number;
  as_of_date: Date;
  as_of_date_report: string;
  due_at: Date;
  closed_at?: any;
  completed_at: Date;
  retired_at?: any;
  strategy_id: number;
  strategy_name: string;
  question_count: number;
  answer_count: number;
  percentage_completed: number;
  is_internal: boolean;
  status_date?: any;
  status_user?: any;
  total_score?: any;
  total_rating?: any;
  primary_contact?: any;
  last_updated_at: Date;
  last_viewed_at: Date;
  view_count: number;
  unauthorized: boolean;
  status_reason: number;
  decline_option: number;
  email_text?: any;
  is_autofillable: boolean;
  linked_id: number;
  created_at: Date;
  created_at_report: string;
  last_reminded_at?: any;
  last_reminded_by_name?: any;
  entity_ids?: any;
  investor_ids?: any;
  all_entity_flag: boolean;
  all_investor_flag: boolean;
  created_by: DiligenceUserType;
  fromfirm_name?: any;
  fromfirm_id: number;
  tofirm_name?: any;
  tofirm_id: number;
  to_user?: any;
  diligence_type: string;
  diligence_reason: number;
  scheduled_for?: any;
  basediligence_id: number;
  parent_entity_id: number;
  shared_count: number;
  linked_duediligence_id?: any;
  presubmission_review_enabled: boolean;
  postsubmission_review_enabled: boolean;
  recalculation_needed: boolean;
  custom_message?: any;
  inbound_configurations_id?: any;
  open_followups_count: number;
  template_firm_id: number;

  // Customs fields added
  alwaysOpen?: boolean;
  isLocked?: boolean;
  isCompleted?: boolean;
  hasReadOnlyAccess?: boolean;
  isReadOnly?: boolean;
  review_allowed?: boolean;
  notVisible?: boolean;
  allowOnlyFollowups?: boolean;
  isQuickFilter?: boolean;
  myFunction: any;
  editor_version: number;
}